requirements.txt
dj-database-url==0.3.0
Django>=1.6,<1.7
Pillow==2.7.0
South>=1.0.2

django-autoslug==1.7.2
django-classy-tags==0.5.2
django-compressor==1.4
django-contrib-comments==1.5
django-cors-headers==1.0.0
django-extensions==1.4.9
django-guardian==1.2.5
django-tastypie==0.12.1
django-userena==1.4.0

psycopg2==2.5.4
python-dateutil==2.1
django-taggit==0.12.2
django-multiuploader==0.2.9
simplejson==3.6.5
django-sendfile==0.3.6
django-haystack==2.3.1
elasticsearch==1.4.0
jsonfield==1.0.0

django-reversion==1.8.5
pygeocoder==1.2.5

sorl-thumbnail==12.2

# vlille
-e git+https://github.com/glibersat/python-vlille.git#egg=python-vlille
